# swagger.json
{"openapi":"3.0.0","paths":{"/cats":{"post":{"operationId":"CatsController_create","summary":"Create cat","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCatDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cat"}}}},"403":{"description":"Forbidden."}},"tags":["cats"]}},"/cats/{id}":{"get":{"operationId":"CatsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cat"}}}}},"tags":["cats"]}},"/file":{"post":{"operationId":"FilesController_uploadFile","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["files"]}},"/roundreplays/{gameId}/{roundNumber}":{"get":{"operationId":"RoundReplaysController_getRoundReplay","parameters":[{"name":"gameId","required":true,"in":"path","schema":{"type":"string"}},{"name":"roundNumber","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoundReplay"}}}}},"tags":["roundreplays"]}}},"info":{"title":"Cats example","description":"The cats API description","version":"1.0","contact":{}},"tags":[{"name":"cats","description":""},{"name":"roundreplays","description":""},{"name":"files","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateCatDto":{"type":"object","properties":{"name":{"type":"string"},"age":{"type":"number"},"breed":{"type":"string"}},"required":["name","age","breed"]},"Cat":{"type":"object","properties":{"age":{"type":"number","example":1,"description":"The age of the Cat"},"breed":{"type":"string","example":"Maine Coon","description":"The breed of the Cat"},"name":{"type":"string","description":"The name of the Cat","example":"Kitty"}},"required":["age","breed","name"]},"Position":{"type":"object","properties":{"tick":{"type":"number"},"x":{"type":"number"},"y":{"type":"number"},"z":{"type":"number"}},"required":["tick","x","y","z"]},"BombEvent":{"type":"object","properties":{"type":{"type":"object"},"tick":{"type":"number"},"userId":{"type":"number"},"position":{"$ref":"#/components/schemas/Position"}},"required":["type","tick","userId","position"]},"Inventory":{"type":"object","properties":{"inventory":{"default":[],"type":"array","items":{"type":"string"}},"userId":{"type":"number"},"tick":{"type":"number"}},"required":["inventory","userId","tick"]},"Kill":{"type":"object","properties":{"tick":{"type":"number"},"killer":{"type":"number"},"assister":{"type":"number"},"victim":{"type":"number"},"weapon":{"type":"string"}},"required":["tick","killer","assister","victim","weapon"]},"BlindedPlayer":{"type":"object","properties":{"playerId":{"type":"number"},"startTick":{"type":"number"},"endTick":{"type":"number"}},"required":["playerId","startTick","endTick"]},"Utility":{"type":"object","properties":{"playerId":{"type":"number"},"type":{"type":"object"},"tickThrown":{"type":"number"},"tickDetonated":{"type":"number"},"tickExpired":{"type":"number"},"throwFrom":{"$ref":"#/components/schemas/Position"},"throwTo":{"$ref":"#/components/schemas/Position"},"damagedPlayerIds":{"default":[],"type":"array","items":{"type":"string"}},"blindedPlayers":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/BlindedPlayer"}},"path":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/Position"}},"entityId":{"type":"number"}},"required":["playerId","type","tickThrown","tickDetonated","tickExpired","throwFrom","throwTo","damagedPlayerIds","blindedPlayers","path","entityId"]},"PlayerHurt":{"type":"object","properties":{"tick":{"type":"number"},"targetPlayerId":{"type":"number"},"healthAfterDamage":{"type":"number"}},"required":["tick","targetPlayerId","healthAfterDamage"]},"PlayerShot":{"type":"object","properties":{"tick":{"type":"number"},"x":{"type":"number"},"y":{"type":"number"},"z":{"type":"number"},"firingWeapon":{"type":"string"},"playerId":{"type":"number"},"yaw":{"type":"number"}},"required":["tick","x","y","z","playerId","yaw"]},"PlayerPosition":{"type":"object","properties":{"tick":{"type":"number"},"x":{"type":"number"},"y":{"type":"number"},"z":{"type":"number"},"playerId":{"type":"number"},"yaw":{"type":"number"}},"required":["tick","x","y","z","playerId","yaw"]},"RoundReplay":{"type":"object","properties":{"roundNumber":{"type":"number"},"startTick":{"type":"number"},"endTick":{"type":"number"},"positions":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/PlayerPosition"}},"playerShot":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/PlayerShot"}},"playersHurt":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/PlayerHurt"}},"utilities":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/Utility"}},"kills":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/Kill"}},"inventories":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/Inventory"}},"bombEvents":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/BombEvent"}}},"required":["roundNumber","startTick","endTick","positions","playerShot","playersHurt","utilities","kills","inventories","bombEvents"]}}}}